# [foreach] Speed up slow path perf for foreach binary ops