@janeyx99 2023-10-11
I am abandoning this change because type promotion is a real concern. Since binary ops are a broader category than just our happy add/sub/mul/div, I don't want to make assumptions about what the Scalar _should_ promote to too early. e.g., something like `eq` will promote types differently than `add`.

Anyway, the approach I will take instead is:
1. Add a scalar Tensor overload for _foreach_add
2. Change the call sites within our core optimizers to be more performant